+++ R/pin_reactive.R
@@ -4,8 +4,8 @@ shiny_dependencies <- function() {
   )
 }
 
-pin_changed_time <- function(name, board) {
-  pin_path <- board_pin_get(board_get(board), name)
+pin_changed_time <- function(name, board, extract) {
+  pin_path <- board_pin_get(board_get(board), name, extract = extract)
   pin_files <- file.path(pin_path, dir(pin_path))
 
   max(file.info(pin_files)[,"mtime"])
@@ -24,9 +24,11 @@ pin_changed_time <- function(name, board) {
 #'   a numeric value.
 #' @param session The user session to associate this file reader with, or NULL if
 #'   none. If non-null, the reader will automatically stop when the session ends.
+#' @param extract Should compressed files be extracted? Each board defines the
+#'   deefault behavior.
 #'
 #' @export
-pin_reactive <- function(name, board, interval = 5000, session = NULL) {
+pin_reactive <- function(name, board, interval = 5000, session = NULL, extract = NULL) {
   deps <- shiny_dependencies()
 
   board_object <- board_get(board)
@@ -35,12 +37,12 @@ pin_reactive <- function(name, board, interval = 5000, session = NULL) {
     intervalMillis = interval,
     session = session,
     checkFunc = function() {
-      changed_time <- pin_changed_time(name, board)
+      changed_time <- pin_changed_time(name, board, extract = extract)
       pin_log("pin_reactive() change time: ", changed_time)
 
       changed_time
     },
     valueFunc = function() {
-      pin_get(name, board = board)
+      pin_get(name, board = board, extract = extract)
     })
 }